【SYKFPT-1093】
feat: complete v0.3.3 Phase 1 internal governance and convergence

- Architecture: Implemented 7-state ProfileWorker state machine with Backoff and Circuit Breaker.
- Config: Enhanced path_parser with Identifier Locator (key:val), array collapsing, and append (+) mode. Added 'config unset' command.
- Storage: Refactored FileStore to hierarchical directory structure, implemented StoreItem trait for normalization.
- Migration: Added v2_to_v3 silent migrator for seamless storage layout upgrade.
- Stability: Eliminated manual drop(lock) patterns in WorkerManager.
- Verification: Passed all 54 E2E test cases.

+++ cli/cowen/crates/cowen-store/src/sql/sqlite.rs
@@ -438,7 +438,7 @@ impl SqlDriver for SqliteDriver {
 
     async fn delete_dlq_by_id(&self, id: i64) -> CowenResult<()> {
         sqlx::query("DELETE FROM cowen_dlq WHERE id = ?")
-            .bind(id as i32)
+            .bind(id)
             .execute(&self.pool).await
             .map_err(|e| CowenError::Store(e.to_string()))?;
         Ok(())
